Disable default Cobra completion commands (#16540)
The Pulumi CLI provides a customised shell completion command,
`gen-completion`, that can be used to generate scripts that enable tab
completion in shells such as Bash and Zsh. When providing custom
generation commands, Cobra (the library we use to build the Pulumi CLI)
[recommends disabling the default completion
command](https://github.com/spf13/cobra/blob/main/site/content/completions/_index.md#adapting-the-default-completion-command).
This commit rebases @EthanOrlander's work in #7842 on `master` and
disables the default command in the recommended manner.

This is technically a breaking change in the interface of the CLI, but
since we'd (presumably) rather that users were only using the officially
supported completions that Pulumi itself maintains, this feels like us
fixing a bug/the right thing to do.

Fixes #7841
Closes #7842

Co-authored-by: Ethan Orlander <ethanorlander@gmail.com>