Provide installation instructions for the python typechecker (#16539)
When the typechecker option is set, but the selected typechecker is not
installed in the project's virtual environment, the pulumi operation
will fail with an unhelpful message like `mypy failed: exit status 1`.

To make this more actionable, we check if the typechecker is installed,
and if not, we provide installation instructions to remediate the issue.