save snapshots after we moved the resources (#16527)
Once we moved the resources, we also want to save the snapshot. Do so
here.

We try to save the destination snapshot first, so resources don't get
orphaned, as it's probably slightly easier for users to delete resources
form the source snapshot than it is for them to import them again in the
destination. Ideally this would be an atomic operation, but I don't
think we have the means to do that right now.