retry post requests that timeout during handshake
We've had quite a few test flakes that happen because of TLS handshake
timeouts to the service backend.  These are usually during stack
creation (which happens a lot during tests), which are POST requests.

POST requests are not generally safe to retry, because the request
could already have been sent to the backend and an action taken, but
the client couldn't read the response.  However when there is a TLS
handshake failure, the request can never have made it to the backend,
so these errors are safe to retry.