Improve TypeDocs for `sdk/nodejs/automation` (#16616)
This commit improves the TypeScript TypeDocs for `automation` modules in
the NodeJS SDK. Specifically:

* It adds documentation to interfaces, properties, etc. that are missing
it.

* It transforms would-be TypeDoc comments erroneously written using
either `/*` (a single asterisk) or `//` (a normal line comment) to
actual TypeDoc comments.

* It standardises on TypeDoc's `{@link Name}` syntax for linking
identifiers, as opposed to the mixture of backticks and square brackets
we have today. This helps IDEs as well as (hopefully) improving any
rendered documentation.

* It fixes typos and generally cleans up the formatting here and there,
as well as introducing more consistency where the same concepts crop up
in multiple places.


![image](https://github.com/pulumi/pulumi/assets/163793/b6ab94cf-74d7-4c6d-a2e1-df737c184794)

![image](https://github.com/pulumi/pulumi/assets/163793/783dd060-67a1-487a-a683-2c8a94ac3d86)