Don't count plugins from a nested policy pack in GetRequiredPlugins
For nodejs we look at the node_modules of all child directories of the
root path to find the required plugins. When a policy pack is nested
within a pulumi project, we were including its dependencies in the list
of required plugins. To avoid this, we stop recursing once we see a
directory that has a PulumiPolicy.yaml.

Fixes https://github.com/pulumi/pulumi/issues/16604