implement engine support for invoke transforms (#16559)
We've recently introduced resource transforms, which allow users to
update any resource options and properties for the duration of a program
using a callback. We want to introduce similar functionality for Invokes
(and eventually also StreamInvokes, Read and Calls). This can help users
e.g. set default providers through transforms consistently for all
components.

While this PR only implements the engine parts of invoke transforms, the
API for this will look very similar to what the API for resource
transforms looks like. For example in TypeScript:

```
pulumi.runtime.registerInvokeTransform(args => {
[...]
});
```

---------

Co-authored-by: Will Jones <will@sacharissa.co.uk>