Install policy or project dependencies based on nearest configuration file (#16631)
When a policy pack is nested within a project, or vice-versa, install
the appropriate dependencies, based on where the command is run from.

If there are both policy pack and project files in the parent
directories of the current working directory, pick the one that is
closest.

Fixes https://github.com/pulumi/pulumi/issues/16605