Test provider `Delete` parameter marshalling
This commit adds a test to ensure that provider `Delete`s are passed
with the correct set of parameters. This is a regression test for #16440,
as part of the follow-up to #16441 in #16484. Assuming we are happy with
this level/style of test, we should consider adding tests for the other
provider methods also.

Fixes #16484