Fix lookup for pulumi run script when using Volta to manage nodejs
We run a short script with nodejs to find the path to the nodejs SDK
entrypoint. In https://github.com/pulumi/pulumi/pull/16160 this script
was changed and used newlines.

This broke when using the Volta package manager to manager different
nodejs versions. Volta adds shim programs into the user's path that
route to the correct nodejs executable. Seemingly Volta does not handle
the newlines in the arguments.

To fix this, we ensure that the script is on a single line.

Fixes https://github.com/pulumi/pulumi/issues/16393