Fix incorrect caching of git auth method in error cases
urlAuthParser.Parse caches auth methods. To avoid caching the auth
method when running into an error, we first checked that the auth method
is not nil. However we ran into a classic Go issue where a nil value of
a concrete type is assigned to a variable of an interface type, making
the it not equal to nil.

https://go.dev/doc/faq#nil_error

https://go.dev/play/p/-xk8tfQr4RA