Default the import base path to something sensible (#16664)
This fixes go package generation so users don't _have_ to set the import
base path explicitly in the go info part of the schema. This only really
helps our packages (because we assume the path is "github.com/pulumi")
but it fixes some of the generated test data and should also work for
local SDKs, where they aren't _really_ at "github.com/pulumi" but also
that it doesn't matter what URL is used as long as it's a valid one.

Fixes https://github.com/pulumi/pulumi/issues/13756

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>