sdk/python: avoid maximum recursion depth error (#16669)
CPython doesn't optimize tail recursion, and because of that has a
fairly low recursion limit set to avoid stack overflows. This also means
that where we rely on recursion we are prone to raising
`RecursionError`s. In this particular case, `_add_dependency` relies on
recursion for finding all the child dependencies.

When deeply nested trees of `ComponentResources` are used,
`_add_dependencies` can end up raising such an error.

Rewrite the function to be iterative instead of recursive to avoid that.

Fixes https://github.com/pulumi/pulumi/issues/16328

(I didn't manage to reproduce the exact stacktrace of the above issue,
so there might be another thing hidden there. But this should help with
it either way)