state_move: show resources to be moved correctly
At some point while moving the code around we lost the output for the
resources to be moved.  We would look for them in the source snapshot
for correct ordering, but the code happened after the URN renaming, so
the resources would not match the ones in the source snapshot anymore.

Fix this by using the resourcesToMoveOrdered list to print the
resources, and also moving this code a bit earlier, as we want to show
the resource URNs from the source snapshot, instead of the rewritten ones.