state_move: handle empty stacks correctly (#16676)
Currently when we have stacks with no snapshot, `pulumi state move`
fails because it tries to use a nil pointer. Handle this scenario
correctly by:
- erroring out if there is no snapshot in the source stack. In this case
there are no resources that can be moved, so there's nothing more that
we can do other than showing the error.
- creating a snapshot in the destination stack if necessary. It's valid
to move a resource to a currently empty stack, so we'll make this work.