state_move: warn no resources are matched/error out when no resources move (#16673)
Currently when no resources are moved, we just show no resources being
moved, but let the command continue. This can be quite confusing to the
user. Error out in that case instead. Also when an argument doesn't
match any resources in the source snapshot, we now warn the user about
that.