Support `--remove` for `destroy` in the Go, NodeJS and Python Automation API SDKs (#16674)
By default, `pulumi destroy` removes all resources within a stack but
leaves the stack and its configuration intact. If one passes the
`--remove` option to `destroy`, however, the stack and its configuration
will also be removed once the resources within the stack have been
deleted. This commit updates the work of @Moon1706 in #11080 to add
`remove` as an option to the Go, NodeJS and Python Automation API SDKs'
`destroy` methods, which then perform an analogous clean-up.

Closes #11080

---------

Co-authored-by: Nikita Sharaev <n.p.sharaev@tinkoff.ru>