avoid duplicate error output in some cases
Cobra has its own error reporting when a command returns an error.
However we duplicate that in `main`.

Usually we don't notice this, because we have our own error reporter
that exits the program early.  However in some cases, such as when the
command is misspelled, that error reporter doesn't kick in, and we end
up with a duplicated error output.

We can just rely on the cobra one in that case, instead of duplicating
that.