Allow specifying project names with `pulumi convert`
When converting a project from one language to another, `pulumi convert`
will default to producing a project named after the source directory. So
if one converts a project in `some_directory`, the resulting
`Pulumi.yaml` will specify `name: some_directory`. This is generally the
right behaviour, since conversion often implies that the target will
supplant the source (and thus should use the same name). However, there
are cases where it can be useful to override this default. This commit
takes @FonsecaGoncalo's work in #15034 and extends `pulumi convert` so
that it accepts an optional `--name` argument, which if supplied (and
valid) will be used instead. As well as rebasing this work, we add a
couple of tests to make sure that both cases work.

Fixes #14645
Closes #15034

Co-authored-by: Gonçalo Fonseca <goncalofnsc@gmail.com>