state_move: reparent provider correctly
Currently when a providers Parent is not moved, we just try to rewrite
the current one using the regular URN rewrite mechanism.  This is not
correct.

Instead, we need to make sure to do the same thing as we do for
regular resources.  Namely if the parent is not being moved into the
destination, we want the parent to be the root stack of the
destination.