improve UI output (#16723)
Change the output for resources to be moved from

```
Planning to move the following resources from dev to dev2:
  urn:pulumi:dev::aws-ts::pulumi:providers:aws::Provider_A
  urn:pulumi:dev::aws-ts::aws:iam/role:Role::Role_A

Do you want to perform this move?  [Use arrows to move, type to filter]
  yes
> no
```

to

```
Planning to move the following resources from dev to dev2:

 - urn:pulumi:dev::aws-ts::pulumi:providers:aws::Provider_A
 - urn:pulumi:dev::aws-ts::aws:iam/role:Role::Role_A

Do you want to perform this move?  [Use arrows to move, type to filter]
  yes
> no
```

for better readability.