state_move: disallow provider moves (#16722)
The `state move` command is designed to move resources with an
underlying physical resource. Otherwise it's easy enough to just
re-create a resource in the resulting program. Especially for providers,
we usually don't want to move them, but rather want to copy them over.
Since there is not much point in moving a provider, just disallow that
and error out if a user tries to do that.