Exclude `pulumi_policy` from plugin determination (#16727)
In the Python language host, we hardcode that `pulumi-policy` doesn't
have an associated resource (provider) plugin, because we know it
doesn't have one.

However, this hardcode no longer works with the latest version of
`pulumi-policy` (v1.11.0) because it was built with a newer version of
`setuptools` which has a behavior change where the package name in the
metadata will now allow underscores, instead of having underscores
replaced with hyphens (https://github.com/pypa/setuptools/pull/4159).
This means that the package name reported from `pip list` is now
`pulumi_policy` instead of `pulumi-policy`, which doesn't match the
hardcoded list.

Note that this change is really only to help with `pulumi-policy`
v1.11.0. Future versions of `pulumi-policy` will have a
`pulumi-plugin.json` file in the package, which properly indicates that
it doesn't have an associated plugin.

Related: https://github.com/pulumi/pulumi-policy/pull/358
Part of addressing: https://github.com/pulumi/pulumi-policy/issues/356