Fix panics due to different length `ignoreChanges` arrays (#16742)
The `ignoreChanges` resource option accepts a list of paths into a
resource that should be ignored when computing whether or not something
has changed. For example:

```typescript
const r = new Resource(
  "r",
  {
    a: "a",
    b: [1, 2],
    c: { d: "d" },
  },
  {
    ignoreChanges: [
      "a",
      "b[*]",
    ],
  }
)
```

Here, when diffing `r`, Pulumi will ignore changes to `a` (due to the
path `"a"`) as well as changes to any element of `b` (due to the
_wildcard_ path `"b[*]"`). Under the hood, `ignoreChanges` is
implemented partly by "resetting" pieces of a resource's state to older
values -- that is, rather than ignoring changes that might be reported,
Pulumi will undo the changes before they are used altogether.

In #16406, a panic encountered when resetting children of arrays of
different length was fixed. This commit extends this fix to cover panics
that occur when simply resetting arrays themselves (e.g. resetting
`[1,2]` to `[1,2,3]`). As part of this, some test cases are renamed to
convey that the are actually testing the children of arrays or
wildcard-selected objects, in order to accommodate new test cases for
this panic.

Fixes #16724