state_move: don't conflict if destination provider with same URN and ID exists (#16737)
When a provider with the same URN and ID already exists in the
destination snapshot, we can assume it's the provider that has been
copied over. This is e.g. useful when running the state move command
multiple times with different resources that have the same provider.

This also allows moving resources back and forth between stacks.