resource_refs_get_resources: upgrade pulumi SDK
The ResourceRefsGetResourcesGo test is flaky because the
`UnsafeAwaitOutput` we're using in it sometimes returns a result where
the value is `nil`, even though it claims it to be known.  We're using
a very old version of the Go SDK for this test, and just upgrading it
fixes the issue.  The await method has indeed changed in the new Go
SDK, so it can be assumed that this flakyness has been fixed in the
meantime, and we should just pull in the new SDK to also unflake the
test.