Add a test that all our conformance test programs are bindable (#16770)
When I was trying to add a new conformance test I had a typo in the PCL
that I didn't pick up till trying to actually run against the NodeJS
host and runtime. This test ensures we won't try to ship a test to the
runtimes that isn't even bindable.