state_move: don't move root stack explicitly (#16779)
The root stack is a special resource, in that it's supposed to exist in
every snapshot. Moving it as we do other resources (and thus
re-parenting it on the existing root stack) doesn't quite make sense.

However it does make sense for moving it to mean "move all the children
of the root stack", and moving it thusly. We can easily achieve this by
just removing the rootstack from the resources to be moved, as it
already exists in the destination snapshot, and all the children will be
re-parented correctly to the destination root stack.

Fixes https://github.com/pulumi/pulumi/issues/16778