[python/automation] Implement Stack.import_resources() for batch importing resources into a stack (#16668)
Addressing https://github.com/pulumi/pulumi/issues/8237 for Python. 

Since `import` is a reserved keyword in python, I went with
`import_resources` as the function name. It's a bit unfortunate that it
is not the same name as other SDKs but I wasn't sure which name to use
here. Suggestions are welcome 🙏